@@ -11,11 +11,11 @@
 		"eslint": "^8.9.0",
 		"eslint-config-haykam": "^1.15.0",
 		"eslint-plugin-react": "^7.23.2",
-		"file-loader": "^4.1.0",
+		"file-loader": "^6.2.0",
 		"html-minify-loader": "^1.4.0",
 		"jsx-loader": "^0.13.2",
-		"webpack": "^4.37.0",
-		"webpack-cli": "^3.3.6"
+		"webpack": "^5.74.0",
+		"webpack-cli": "^4.10.0"
 	},
 	"scripts": {
 		"build": "webpack --config webpack.config.js",
@@ -24,9 +24,9 @@
 	"dependencies": {
 		"debug": "^4.1.1",
 		"prop-types": "^15.7.2",
-		"react": "^16.8.6",
-		"react-dom": "^16.8.6",
-		"styled-components": "^4.4.1"
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"styled-components": "^5.3.6"
 	},
 	"eslintConfig": {
 		"extends": [
