@@ -1,12 +1,9 @@
 require("file-loader?name=[name].[ext]!html-minify-loader!./index.html");
 
 const React = require("react");
-const ReactDOM = require("react-dom");
+const { createRoot } = require("react-dom/client");
 
 const App = require("./components/app.jsx");
 
-const { main: log } = require("./util/debug.js");
-
-ReactDOM.render(<App />, document.getElementById("app"), () => {
-	log("rendered app");
-});
+const root = createRoot(document.getElementById("app"));
+root.render(<App />);
